Reorganise configuration struct (#12)
